finishing db stuff
